requirements.txt
# ── Web framework ─────────────────────────────────────────────────────────────
Django>=4.2,<5.0
python-dotenv>=1.0.0
django-cors-headers>=4.3.0
django-extensions>=3.2.0

# ── LLM providers ─────────────────────────────────────────────────────────────
openai>=1.30.0
anthropic>=0.40.0

# ── ML / retrieval pipeline ───────────────────────────────────────────────────
sentence-transformers>=2.2.0   # bi-encoder + cross-encoder
rank-bm25>=0.2.2               # BM25 lexical retrieval
scikit-learn>=1.3.0            # TF-IDF, cosine similarity
numpy>=1.24.0

# ── Document extraction (all resume formats) ──────────────────────────────────
pypdf>=4.0.0                   # PDF
python-docx>=1.0.0             # DOCX
beautifulsoup4>=4.12.0         # HTML
Pillow>=10.0.0                 # image pre-processing
pytesseract>=0.3.10            # OCR wrapper (needs: brew install tesseract)

# ── Optional: MySQL database ──────────────────────────────────────────────────
# mysqlclient>=2.2.0           # uncomment if USE_SQLITE=false in .env
